+++ blueprint/interfaces/worker/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @forklaunch/interfaces-worker
 
+## 1.0.3
+
+### Patch Changes
+
+- Worker decryption and encryption update
+
 ## 1.0.2
 
 ### Patch Changes